+++ src/modules/angular-formly-templates-lumx.js
@@ -2,6 +2,12 @@
 	'use strict';
 	var USING_TEMPLATES = true;
 	var USING_TEMPLATE_VALIDATION = true;
+	/* Custom validation message defaults here */
+	var VALIDATION_MESSAGES = [{name: 'required', message: 'This field is required'}, {
+		name: 'maxlength',
+		message: 'This field is too long.'
+	}, {name: 'minlength', message: 'This field is too short.'}];
+	/* Module Templates + Data */
 	var MODULE_NAME = "formlyLumx";
 	var PREFIX = "lx";
 	var FIELDS = [{
@@ -71,8 +77,8 @@
 		}
 	}
 
-	function addFieldValidationOptions(apiCheck) {
-		var VALIDATION_FIELDS = [{
+	function addFieldValidationOptions(apiCheck) {  /* validation using apiCheck.js */
+		var APICHECK_VALIDATION_FIELDS = [{
 			"name": "checkbox",
 			"validateOptions": {
 				"label": apiCheck.string,
@@ -141,7 +147,7 @@
 				"cols": apiCheck.number
 			}
 		}];
-		VALIDATION_FIELDS.map(function (validationField) {
+		APICHECK_VALIDATION_FIELDS.map(function (validationField) {
 			FIELDS.map(function (field) {
 				if (field.name === validationField.name) {
 					field.validateOptions = validationField.validateOptions;
@@ -173,6 +179,16 @@
 		}
 	}
 
+	function setDefaults(formlyConfig, formlyValidationMessages) {
+		formlyConfig.extras.ngModelAttrsManipulatorPreferBound = true;
+		VALIDATION_MESSAGES.map(function (validation) {
+			formlyValidationMessages.addStringMessage(validation.name, validation.message);
+		});
+		formlyValidationMessages.messages.pattern = function (viewValue, modelValue, scope) {
+			return (viewValue || modelValue) + ' is invalid';
+		};
+	}
+
 	function cacheTemplates($templateCache) {
 		if (USING_TEMPLATES) {
 			FIELDS.map(function (field) {
@@ -184,5 +200,5 @@
 		}
 	}
 
-	angular.module(MODULE_NAME, ['formly']).config(setWrappers).run(setFields).run(cacheTemplates);
+	angular.module(MODULE_NAME, ['formly']).config(setWrappers).run(setFields).run(setDefaults).run(cacheTemplates);
 }());